+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added redirect to remote_accounts#create when no remote account is present.
+* Added filtering of rental lists.
 
 # 0.0.1