+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 0.0.2
+
 * Added redirect to remote_accounts#create when no remote account is present.
 * Added filtering of rental lists.